refactor: rename DataReader as S3Interpreter
