+++ locations.json
@@ -1,1 +1,52 @@
-[{"x":911.1647857041422,"y":378.23135333002955,"id":"dceb95ed-92b4-4a53-98aa-0d59b7494360"},{"x":967.0915228878728,"y":859.4645637731402,"id":"336981e7-8813-4fa6-802d-581186057981"},{"x":987.6197623631176,"y":75.42606807163321,"id":"0364a4a7-b0bd-46db-a8a4-0657ed497e46"},{"x":633.1289212645819,"y":156.3366389039713,"id":"b4e9df06-da1a-4523-bed8-63049fcf85e0"},{"x":323.3469707232768,"y":9.752536538759514,"id":"981ee9c8-e828-47d4-a673-de8b57d3d57d"},{"x":93.86391895257329,"y":922.8832854435598,"id":"26cedb64-4b2c-4c63-b9b2-f426e56e6056"},{"x":444.0111430710174,"y":945.4809657947889,"id":"9560bfd0-64ce-4980-85de-b9208b982a65"},{"x":27.377301425072332,"y":279.0555889270385,"id":"799a0fa5-fe44-4c77-999d-536692b7efc7"},{"x":637.761402823104,"y":607.074189123254,"id":"8d472794-00e6-4c7a-b867-6fd596b42aaa"},{"x":342.05883061909867,"y":322.9135261375742,"id":"3ef1a983-7f31-48bb-93bd-02d25b949916"}]
+[
+  {
+    "x": 911.1647857041422,
+    "y": 378.23135333002955,
+    "id": "dceb95ed-92b4-4a53-98aa-0d59b7494360"
+  },
+  {
+    "x": 967.0915228878728,
+    "y": 859.4645637731402,
+    "id": "336981e7-8813-4fa6-802d-581186057981"
+  },
+  {
+    "x": 987.6197623631176,
+    "y": 75.42606807163321,
+    "id": "0364a4a7-b0bd-46db-a8a4-0657ed497e46"
+  },
+  {
+    "x": 633.1289212645819,
+    "y": 156.3366389039713,
+    "id": "b4e9df06-da1a-4523-bed8-63049fcf85e0"
+  },
+  {
+    "x": 323.3469707232768,
+    "y": 9.752536538759514,
+    "id": "981ee9c8-e828-47d4-a673-de8b57d3d57d"
+  },
+  {
+    "x": 93.86391895257329,
+    "y": 922.8832854435598,
+    "id": "26cedb64-4b2c-4c63-b9b2-f426e56e6056"
+  },
+  {
+    "x": 444.0111430710174,
+    "y": 945.4809657947889,
+    "id": "9560bfd0-64ce-4980-85de-b9208b982a65"
+  },
+  {
+    "x": 27.377301425072332,
+    "y": 279.0555889270385,
+    "id": "799a0fa5-fe44-4c77-999d-536692b7efc7"
+  },
+  {
+    "x": 637.761402823104,
+    "y": 607.074189123254,
+    "id": "8d472794-00e6-4c7a-b867-6fd596b42aaa"
+  },
+  {
+    "x": 342.05883061909867,
+    "y": 322.9135261375742,
+    "id": "3ef1a983-7f31-48bb-93bd-02d25b949916"
+  }
+]
