+++ src/main/java/eventprocessing/GlobalConstants.java
@@ -1,0 +1,4 @@
+package eventprocessing;
+
+public class GlobalConstants {
+}
