+++ src/main/java/eventprocessing/fileservices/S3Interpreter.java
@@ -1,4 +1,4 @@
-package eventprocessing;
+package eventprocessing.fileservices;
 
 import com.amazonaws.services.s3.model.S3ObjectInputStream;
 
@@ -6,7 +6,7 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 
-public class DataReader {
+public class S3Interpreter {
     public void saveToFile(S3ObjectInputStream amazonS3Stream) throws IOException {
         FileOutputStream fos = new FileOutputStream(new File("locations.json"));
         byte[] read_buf = new byte[1024];
